Get named Oxford colours vector

Get named Oxford colours vector

Usage

get_oxford_colour(pattern = NULL, type = c("hex", "rgb", "cmyk", "pantone"))

get_oxford_colours(pattern = NULL, type = c("hex", "rgb", "cmyk", "pantone"))

Arguments

pattern

Optional. A character value or vector to use as a search term. Default is NULL in which case all the Oxford colours are returned.

type

A character vector of colour type. Can be "rgb", "cmyk", "hex", or "pantone". Default is "hex".

Value

A named vector of Oxford colour/s

Examples

get_oxford_colours()
#>          Oxford blue         Oxford mauve         Oxford peach 
#>            "#002147"            "#776885"            "#E08D79" 
#>  Oxford potters pink          Oxford dusk         Oxford lilac 
#>            "#ED9390"            "#C4A29E"            "#D1BDD5" 
#>        Oxford sienna       Oxford CCB red          Oxford plum 
#>            "#994636"            "#AA1A2D"            "#7F055F" 
#>         Oxford coral      Oxford lavender        Oxford orange 
#>            "#FE615A"            "#D4CDF4"            "#FB5607" 
#>          Oxford pink         Oxford green    Oxford ocean grey 
#>            "#E6007E"            "#426A5A"            "#789E9E" 
#>  Oxford yellow ochre     Oxford cool grey      Oxford sky blue 
#>            "#E2C044"            "#E4F0EF"            "#B9D6F2" 
#>    Oxford sage green      Oxford viridian    Oxford royal blue 
#>            "#A0AF84"            "#15616D"            "#1D42A6" 
#>          Oxford aqua   Oxford vivid green    Oxford lime green 
#>            "#00AAB4"            "#65E5AE"            "#95C11F" 
#> Oxford cerulean blue  Oxford lemon yellow      Oxford charcoal 
#>            "#49B6FF"            "#F7EF66"            "#211D1C" 
#>      Oxford ash grey         Oxford umber    Oxford stone grey 
#>            "#61615F"            "#89827A"            "#D9D8D6" 
#>    Oxford shell grey     Oxford off white                 Gold 
#>            "#F1EEE9"            "#F2F0F0"                   "" 
#>               Silver 
#>                   "" 
get_oxford_colours(type = "rgb")
#>          Oxford blue         Oxford mauve         Oxford peach 
#>          "0, 33, 71"      "119, 104, 133"      "224, 141, 121" 
#>  Oxford potters pink          Oxford dusk         Oxford lilac 
#>      "237, 147, 144"      "196, 162, 158"      "209, 189, 213" 
#>        Oxford sienna       Oxford CCB red          Oxford plum 
#>        "153, 70, 54"         "127, 5, 95"         "127, 5, 95" 
#>         Oxford coral      Oxford lavender        Oxford orange 
#>        "254, 97, 90"      "212, 205, 244"         "251, 86, 7" 
#>          Oxford pink         Oxford green    Oxford ocean grey 
#>        "230, 0, 126"        "66, 106, 90"      "120, 158, 158" 
#>  Oxford yellow ochre     Oxford cool grey      Oxford sky blue 
#>       "226, 192, 68"      "228, 240, 239"      "185, 214, 242" 
#>    Oxford sage green      Oxford viridian    Oxford royal blue 
#>      "160, 175, 132"        "21, 97, 109"        "29, 66, 166" 
#>          Oxford aqua   Oxford vivid green    Oxford lime green 
#>        "0, 170, 180"      "101, 229, 174"       "149, 193, 31" 
#> Oxford cerulean blue  Oxford lemon yellow      Oxford charcoal 
#>       "73, 182, 255"      "247, 239, 102"         "33, 29, 28" 
#>      Oxford ash grey         Oxford umber    Oxford stone grey 
#>         "97, 97, 95"      "137, 130, 122"      "217, 216, 214" 
#>    Oxford shell grey     Oxford off white                 Gold 
#>      "241, 238, 233"      "242, 240, 240"                   "" 
#>               Silver 
#>                   "" 
get_oxford_colours(pattern = "lilac")
#> Oxford lilac 
#>    "#D1BDD5" 
get_oxford_colours(pattern = c("lilac", "sage green"))
#>      Oxford lilac Oxford sage green 
#>         "#D1BDD5"         "#A0AF84"